Migrate download status rule
